feat: WITH|WITHOUT PARENT TRACKER
You can now search for artifacts that are children of an artifact
belonging to a named tracker:
`WITH PARENT TRACKER = 'epic'`
=> will return all artifacts that are children of an `epic`, regardless
of the project.

Or the ones that are not children:
`WITHOUT PARENT TRACKER = 'epic'`
=> will return all artifacts that are not children of an `epic`.

Implementation note: since the whole TQL implementation in tracker
reports rely on legacy DAO, I had to use the deprecated method
`quoteSmart`. It would be great to convert it to EasyDB like it is done
for cross tracker, but it should be an independant development.

Part of story #32281: Search artifacts with/without parents via TQL

Change-Id: I8af73cf8fd6346492c0ddf23f9394501fe668740