fix: add csrf protection when register a passkey
Part of story #31175 delete a registered passkey

To avoid csrf attack on POST /webauthn/registration, add a csrf token
in the request body and then check it in controller.

*Testing:*

If you try to call the route without the csrf token (or with a wrong
token) you will receive a 400 Bad Request.
Else registration is performed, and you receive 200 OK.

Change-Id: Ia73a6dd18fe061f2545999e6b85c66c791abc071