Fix request #32288 Custom names for the Program Increment and the Interation are not displayed on the Program administration page
In the Program administration page, we can set a custom label for the program
Increment and the iteration that will be displayed on the program management tool.

After setting the value, when we go back to the Program administration page,
the custom labels that we defined are not displayed. If we do other modifications
(for example add a new tracker on the backlog) and we click on
"Save milestone configuration", we will loose all the custom labels.

How to test
-----------

* Go the administration page of a program
* Set custom labels for iterations and programs
* Refresh the page

=> Custom labels must be set in the form

Change-Id: I8846709bfb28fe9b0850651d35c6c3a7c3ccd91d