fix: add csrf protection when delete a passkey
Closes story #31175 delete a registered passkey

To avoid csrf attack on DELETE /webauthn/key/{key_id}. Change endpoint
to POST /webauthn/key/delete and add a csrf token in the request body.
The key_id is also now in the body.

*Testing:*

Try call the route without the csrf token (or with a wrong one) you will
receive a 400 Bad Request.
Else deletion is performed and you receive 200 OK.

Change-Id: Ib2fd06a0cb2f557af08e9271321cc884b07ed7a0