feat: add webauthn lib for strong authentication
Part of story #31177 check my identity when doing a risky action

This commit introduces a global ts lib which bring one function:
authenticate. This function performs an authentication along WebAuthn
protocol and returns if it succeed or not in a ResultAsync object. The
Fault is wrapped in Err if not succeed.

Already use this library for the Check button in Webauthn user
preferences.

*Testing:*

Just call the function as it is:
- If your browser doesn't support WebAuthn, it returns Ok
- If you have no registered passkeys, it returns Ok
- If authentication failed, ... it returns Err ;)
- Else it returns Ok.

You can also just click on the button mentionned above.

Change-Id: I92d61ad2804cd762bf8aa7abe42ab1337396187e