feat: Create a fake new link
How to test:
- In your web container, enable the feature flag: `tuleap config-set
feature_flag_create_artifact 1`
- Edit an artifact with an artifact link field with the modal (kanban,
TTM or planning view)
- Click on the "Create a new artifact" button in the link selector
dropdown.
- Write a title
- Click on the "Create" button. It will close the artifact creation
section and add a new link to the links field. The contents of the new
link are hardcoded (the title is ignored) for now. The id is voluntarily
engineered to cause a Bad Request error in case you submit the modal, to
prevent data corruption.
- If you leave the title blank, the browser should stop the submit and
warn you that it is required.

part of story #28823 [modal] create artifact

Change-Id: Ic1b7e2f91e44a5930c125325eae611facce3370d