feat: Select current tracker at init and memorize tracker choice
How to test:
- In your web container, enable the feature flag: `tuleap config-set
feature_flag_create_artifact 1`
- Edit an artifact with an artifact link field with the modal (kanban,
TTM or planning view)
- Click on the "Create a new artifact" button in the link selector
dropdown.
- The current tracker should be selected in the trackers list. It can be
saved as selected despite being disabled, this will be fixed on a
separate contribution.
- If you select another tracker, it is memorized: if you close the
creation section and reopen it (without closing the whole modal), it
will be remembered.
- Write a title and click on the "Create" button. It will add a new link
with your title. The rest of the data is still hardcoded for now.

Notes:
The Trackers selectbox is _not_ required on purpose. Making it required
will force the browser to select the first possible value if the current
tracker is disabled, or if the selected tracker is no longer in the
list. We will write custom validation on a separate contribution to
prevent people from submitting when the selected tracker is Nothing.

part of story #28823 [modal] create artifact

Change-Id: I1cd53f5bb5c659ed96f6e00b646d109d5bbc9fd8