Closes request #32295: Do not activate all installed plugins in the Docker images by default
Not all plugins are installed out of the box when the containers are
first started. A specific list is now defined.
Only fresh install are affected by this change, for existing instances
the state of the plugins is not modified.

For TEE images both FTS DB and FTS Meilisearch package are installed in
the image. The FTS DB plugin is enabled by default as it does not
require an additional deployment.

Change-Id: I7b56616aa74063b515ccd19ccb1394edc48f5fc0