feat: Use webauthn lib for ssh keys add button
Part of story #31177 check my identity when doing a risky action

Adding a ssh key to an account is a potential vector of attack. So
before doing it, check identity of user with its passkey

*Testing:*

Try to add a ssh key to your account, the behavior of the button in the
modal is the same as before except that if you have a registered passkey
you will be asked to used it.

Change-Id: I7a300fe7c18d91e771c09e948dee0b0e83bde3d0