Prepare release of @tuleap/project-sidebar 2.2.0
8ee4c6ae7bacda527fbdf7e0a79c6632da887366 made changes
that are now missing in the released sidebar component.
A new release is needed because icons are now missing
most notably the icon for the tracker service.

@fortawesome/fontawesome-free has been upgraded from 6.1.1 to 6.4.0
to avoid needing re-releasing the component soon. See changelog here:
https://fontawesome.com/docs/changelog/

The deps of of the package have been switched back to the link: protocol,
using the workspace: protocol with `pnpm publish` in our current setup does
not work.

Part of request #32284: Homogenize TLP icons

Change-Id: I27c5332da90d4f935098e88c2f41e905da56bbcd