Disable MediaWiki parser cache
This cache was disabled in our legacy MW (setting was named `wgEnableParserCache`).
This cache is interesting for large farm, in our case with multiple
small wiki and not that much traffic the gain are not substantial and it
might be a source of issues.

To test, build and redeploy the package on your dev instance.

Part of request #32298: Not the same content between view & edit pages

Change-Id: I1154776e77a86de0344e5853cdfb11076d227fd8