feat: WebAuthn lib use a modal to advertise user on authentication
Part of story #31177 check my identity when doing a risky action

To improve user experience, when doing a risky action, the click on the
button open a first modal to advertise user about identity check. If it
succeeded this modal is closed and the 'real' modal is opened.

*Testing:*

The lib is already used for ssh keys and access keys. If you have no
registered passkeys, you will see no differences. But if you have at
least one registered passkey the authentication modal should appear.

Co-authored-by: Joris Masson <joris.masson@enalean.com>
Change-Id: I4c45ab3a6ca5e5900eb5ca08ef64e24c271f718b