fix: cypress test fail on ssh and access keys
Part of story #31177 check my identity when doing a risky action

Cypress test fail because it want to click on button to open modal
before modal created. The modal was not yet created because it fetch
backend before and cypress goes to fast. To fix that, when fetching the
button is disabled.

Change-Id: I582807c7d9cb1f6460f89c548d5ca5f9ca134f67