feat: Search in list of projects
How to test:
- In your web container, enable the feature flag: `tuleap config-set
feature_flag_create_artifact 1`
- Edit an artifact with an artifact link field with the modal (kanban,
TTM or planning view)
- Click on the "Create a new artifact" button in the link selector
dropdown.
- The list of projects is a list-picker. You can filter values in it to
search for the project you want.

Notes:
- There is an issue when you clear the project list-picker by clicking
on the "cross" icon to remove all values. The list-picker will appear
empty, but the inner `<select>` will have the value of the first project
of the list. The code will proceed and act like that project is
selected, but the list-picker will appear empty. This issue should be
adressed specifically on a dedicated contribution.

Change-Id: I896d588ccb27cf457b78fbb53938e6d2cab68041