Fixes request #32629: XSS in the card field of the agile dashboard apps
Tried to keep the changes within the scope of the security issue. The
template engine of hybrids is bit abused but it is still better than
before. A future contribution might want to consider removing angular
completely from @tuleap/plugin-agiledashboard-card-fields.

This change also fixes the display of the legacy colors for selectbox
using them.

Change-Id: I301ed6a72c324054a490605e881646bf317316fb