Implement request #32632 getAllJSON should have default collection type parameter
No functional change expected.

Notes:
The order of the generic type parameters is reversed: you now only need
to provide the type of item you expect to get in the returned array. If
the endpoint returns arrays, that's it, everything works. If the
endpoint returns an object shape, you need to specify the shape of the
JSON payload (as well as the `getCollectionCallback`), just like before.

Change-Id: Ia15a1484feb0612400de8b8ed5e7a289a615efcf