Closes request #32633 Meilisearch: 1.0.2 -> 1.2.0
The upgrade process for users of the "local Meilisearch"
server has been automated. As it involves re-indexing
everything some items might not be visible until the re-indexation
process is complete.

To test, rebuild the package and deploy it on dev instance.
If you have already indexed items in a previous version the
data will be cleaned and the re-index process will be started
automatically.

Changelogs:
https://github.com/meilisearch/meilisearch/releases/tag/v1.2.0
https://github.com/meilisearch/meilisearch/releases/tag/v1.1.0

Change-Id: I7cc2ed5f43f2e28df99d8a18cb2775c754ea37e4