refactor: kanban/:id/* to dedicated kanban plugin
Only some part of the REST routes are moved to a dedicated kanban folder.

For now it is not a real plugin. Strategy is to move code bit by bit to
not have 30k+ LOC to review if we move the whole feature at once.

Once the feature will be moved, we will have a better view of
dependencies between A.D and Kanban. At the end we will be able to have
a truly independent plugin.

No functional changes.

Part of request #32645: Extract kanban from agiledashboard

Change-Id: I675307d5fecf2934c8b9cf6c739d9c607b7ed8c9