Correct aria-expanded that have been updated between vue2 and vue3
part of request #31534: vue document should be in vue3

Corrects the following warning
(deprecation ATTR_FALSE_VALUE) Attribute "aria-expanded" with v-bind value `false` will render aria-expanded="false" instead of removing it in Vue 3. To remove the attribute, use `null` or `undefined` instead. If the usage is intended, you can disable the compat behavior and suppress this warning with:

Change-Id: Iebb319b1046f462bc5c076890cc5ad47804b9da2