refactor: regroup getCriteriaFrom and getCriteriaWhere
Since we have parametrized FromWhere at our disposal, we don't need
anymore to have distinctive calls to retrieve the from and the where of
a criterion.

We now have the different SQL fragment at the same place (mostly), this
avoid to go back and forth between the construction of the from and the
where to understand what's going on.

No functional changes

Note: Some issues have been fixed during development of this refactoring
(extra placeholder in Integer query, missing spread operator in Float
😭). This will probably have to be backported.

Closes request #32643: regroup getCriteriaFrom and getCriteriaWhere

Change-Id: Ic48c536e1cb0a1124fb3e043c5736bc6b5b7a33a