refactor: move card-fields lib to cardwall
"@tuleap/plugin-agiledashboard-card-fields" is used by both planning and
kanban. Since we are moving kanban in a dedicated plugin, we have to
move card-fields lib to a dedicated place.

Since the semantic card fields is managed by cardwall plugin, it seems logical
to have the lib card-fields also managed by cardwall.

This forces us to move the build of cardwall.js to be moved in a legacy
folder with an independant build.

BTW the unused Tracker_CardFields class that was coming from planning v1
is deleted.

No functional changes:
* Card fields are still displayed in pv2 and kanban
* Cardwall cards can still be moved/stacked/…

Part of request #32645: Extract kanban from agiledashboard

Change-Id: Ia9d566c882bda0fed98d18ce7046b923c36e4175