Closes request #32651 "Artifact can only have one parent" is confusing
The note is now displayed in a `title` attribute. This title attribute
is displayed in the link type selector when a parent artifact is already
linked.
To see it you have to have an already linked a parent artifact with the current
artifact, then open the link type selector and hover the `Is Child
of` type.
You syhould see the reason why the type is disabled
You can test with the creation mode and the edition mode of the artifact
modal.

Change-Id: I48a2c096959d777d42dba52ce5ca3a5bb38f33f8