feat: No more customisation of kanban title
We intend to move Kanban as a dedicated plugin/service. The name of the
feature should not be customiable anymore (was customisable under a.d »
admin » kanban). Starting from now it will always be "Kanban".

kanban_title attribute in XML exports is ignored.

Doing it now allows us to have less code to move from a plugin to
another.

Part of request #32645: Extract kanban from agiledashboard

Change-Id: I30e9f7186ce11a8e5ad50e4105414000d7ba84fb