feat: Introduce kanban plugin
After `make post-checkout` you still have a working kanban (project
creation, cards manipulation, A.D administration, widgets, …)

If admin deactivate A.D and not kanban, then kanban is considered part
of tracker service as a fallback.

A.D depends on kanban, so we cannot disable kanban if A.D is activated.

Part of request #32645: Extract kanban from agiledashboard

Change-Id: I2d3b181c9909ac665125322c05cee0047a99bd8f