feat: Add site admin panel to see registered key of users
Part of story #31178 see keys of user

This commit introduces a new panel in site administration user details
to look for registered passkeys of users

*Testing:*

If plugin is not enabled, nothing changed. If plugin is enabled, a new
panel appear in bottom right of screen with registered passkeys of the
user.

Change-Id: I66d3df94b0d58716d42f140626136e4a93a19255