Implement request #32650 Artifact link list should display parent first
How to test:
- Edit an artifact with an artifact link field with the modal (kanban,
TTM or planning view).
- If it has a parent, it should be at the top of the list with a thicker
border.
- Multiple parents are handled, the border shows up only once. You can
assign multiple parents through the API or the Artifact view (but not
the modal).
- If you change the type of a parent link, it should not "jump" to the
bottom of the list.
- Marking artifacts for removal works as before.
- If the artifact didn't have a parent and you add a new one, it should
not "jump" to the top of the list. It stays mixed with the other new
links. If you submit and reopen the modal for this artifact, it should
now be sorted at the top of the list.

Change-Id: I58536505f494c8ddeb627d8d939f416583448e6f