chore: Add multi trackers and dedicated tracker terminology in existing code
This is part of story #32788 Introduce the Jira project import with a solo tracker

This patch splits the responsibility of importing all issues form a
Jira project or only issues from a selected Jiura issueType in dedicated objects.
This will ease the introduction of the import of all issues in a solo
tracker.

No functionnal change expected here in all existing imports.

Note: The public methods used to build XML content will be extracted in
dedicated objects in the next commit.

Change-Id: Ic3e4c8d8491c48eb81019cbc610de039ed25c2df