feat: Add context to webauthn modal error message
How to test:
When there is an issue with the POST authentication challenge endpoint
(to test, make it return 400 Bad request, for example), the user
account page will show an error message with a translated message for
context: There was a problem while checking the registered passkeys of
user. Previously, only the error code was shown in the alert (Bad
Request), which was puzzling for the end-user (which request ?)

No other functional change expected in the Keys & Tokens page.

part of story #31177 check my identity when doing a risky action

Change-Id: Ief52ef8dd72a7ca443aecd858852fe2e9a37482c