fix: Add length limit for passkey name
Part of story #31177 check my identity when doing a risky action

In database the passkey name can contains only 255 characters. UI allow
to input as many as we want. This commit adds attribute maxlength on the
input and also check in backend the length of the name

*Testing:*

While registering a passkey, you can't input more than 250 characters.
If you cheat, the backend will return an error 400 with
'Name of passkey can contains 250 characters maximum'.

Change-Id: Ia342b3c60d58c99027590024ab916a33dd5b9fe9