Fixes request #32990: Crash while saving a document edited collaboratively in ONLYOFFICE
We should not expect that constraints imposed by our infrastructure/the
way we store data to be respected outside of the infrastructure data layer.
Especially when those constraint are not that easy to enforce at the
type level and even more when those data are generated by a third party
(in this case ONLYOFFICE).

Not exactly sure how to reproduce the issue as it requires to know
exactly how ONLYOFFICE works. It is likely caused by a scenario like
this one:
1. 2 users start to edit a document in ONLYOFFICE
2. The first user logout from Tuleap while the second user continue to
   work
3. The first user login again and continue to edit the document
   alongside the other user

In any case, the integration tests demonstrate the issue.

Change-Id: I947fe7a1c9ce3f99cf7844b9311d505167c4583a