chore: Unit-test bubbling change event on all fields
No functional change. Unit-tests should pass.

Notes:
Add unit-tests for all mutable fields to make it clear the "change"
event is part of the fields' contract to the modal. When we change those
fields, we need to keep that in mind. For example if we add ShadowDOM to
the fields, the native "bubbling" will stop at the Shadow root. The host
will need to keep dispatching at least one "change" event so that the
modal can show the warning before closing.

Closes request #32653 Do not lose modifications done in Artifact Modal

Change-Id: I96334a0cc2e1e3b5299961ac09f528625cd23fc4