feat: Move ttm steps definition fields
part of story #32280 Complete the "move artifact" feature

This contribution allows ttm's steps definition field to be moved.
Please note that the steps execution field is systematically considered
as a not migrateable field. The main reason is that no XML is written
for it when an artifact is exported.

How to test:
- Have a test case containing steps
- Move it to another test case tracker
--> It works

- Have a test execution artifact
- Move it to another test execution tracker
--> The dry run step says the step results field it won't be moved.

Change-Id: I0893a59d5006dbdaaee64fb83e54cc4e62ee1892