feat: Add parameter to the edition modal callback
How to test:
No functional change in the artifact modal. Callers of "showEdition" are
not broken.

Note:
The parameter is hardcoded for now but will be dynamic in a separate
contribution. The goal of this patch is to set up the groundwork and
check that all JS callers are not broken.

part of request #32986 Planning doesn't refresh when links are updated
with modal

Change-Id: I91ea9986c3cbe2634a138eef3eaa2bc5651934c9