Fix request #32998 Cannot upload a file using the modal if my group cannot submit or edit
Given a tracker in a public project with a file field having the following permissions :

* all_users: Read only
* registered_users: can submit
* project_members: can update

Then a project member can add a file using the modal. However, a regular user not member cannot
add a file at artifact creation using the modal (it works as expected in the artifact UI).

Change-Id: Id4b4dd82fbc7257913f7f85c19de228f665c98db