feat: Reload after editing milestone links
How to test:
- Go to the top backlog of a project with Agile Dashboard
- Edit a Release with the modal. Touch the artifact links field.
- After submitting the modal, the page is reloaded so that added/removed
user stories are correctly shown. A success feedback is shown for 5
seconds.

- Edit a Release with the modal, but do not touch the artifact links
field.
- After submitting the modal, the release is refreshed (the changed info
shows up in Planning view).

Note:
The message is not the same as for user stories because we don't have
the tracker shortname in the submilestone data. Also, the milestones' ID
is not visible in the user interface, so it makes more sense anyway to
use their label.

part of request #32986 Planning doesn't refresh when links are updated
with modal

Change-Id: Idbb7f6c02a670f4d5c19bd2931b309a12ef61840