feat: forbid move when there are no fields to move
part of story #32280 Complete the "move artifact" feature

Note:

On the backend-side, the move is forbidden only for the duck typed move.
Since the behavior already existed for the semantic move, and that this
move type will be removed, I think it can be left like that.

How to test:
- Have a two trackers with incompatible fields
- Pick up an artifact in one of these trackers
- Try to move it to the second tracker
--> The dry run steps says the artifact can't be moved
--> The submit button is disabled

- Using the web console, remove the "disable" attribute from the submit
  button, then click it
--> A rest error says it cannot be moved.

Change-Id: Ia33e844a4856a80fbdd76cc2366c77b457480c40