fix: Registration of passkey with eddsa
Closes story #31177 check my identity when doing a risky action

When you have a yubikey with firmware 5.4.3, it uses eddsa algorithm.
But the response of the passkey is malformed. After have openned an
issue on web-auth/webauthn-framework github repo. The repo owner fixed
this problem with a workaround (https://github.com/web-auth/webauthn-framework/pull/437)

This commit update the dependency to web-auth/webauthn-lib to version
4.6.4.

*Testing:*

If your yubikey was impacted by this problem, you can now register it.
For others no changes expected.

Change-Id: I375164517dfcb0607d2f5d1534130b4024f5c669