feat(jira): introduce feature flag to deal with mono tracker import
This is part of story #32788 Introduce the Jira project import with a solo tracker

The imported mono tracker is empty for now.
The feature flag will be updated to be a CLI option once all is done and
validated.
Having an hidden feature flag will allow us to do incremental
development of the feature.
For now, board and agile data retrieving are disabled in the mono
tracker import

How to test
-----------

* Set the feature flag feature_flag_jira_import_mono_tracker_mode to 1
* Run a Jira project import

=> The project is created with an empty mono tracker without any data.

* Set the feature flag feature_flag_jira_import_mono_tracker_mode to 0
* Run a Jira project import

=> The project is created with an multi trackers and with all expected data (as before).

Change-Id: Ide7c7d89557b849731b25882d8bc0ea6661c2cd2