Drop useless checks when checking the criteria value of the data field
Follow up to 94ec889534a6067665f7aaf1b975642dbdaaca42.
A bunch of checks were made but they do not make a lot of sense.
This change only keep what is necessary. This also opens up the road to a
cleaner refactoring where a proper object is returned instead of an
unspecified array.

request #33316: Cardwall based tracker renderer widget is empty when report filters on date fields in normal mode

Change-Id: I04c79fb446d2878d7c53d4c26ea68229c303a8a1