feat: Add page for user to login passwordless
Part of story #33005 login with only my passkey

Add a button on login page to access to passwordless login and a simple
page for it. This page will perform first part of webauthn
authentication, the second will be handled as a form by a controller.

*Testing:*

On login page (homepage and account/login.php) a passwordless button
appear and on click bring user to another page with an input and a
button actived if input is not empty.

Change-Id: Id236da9a7158f12eb139f4e9be9f274b34c324da