refactor: move homepage styles out of shared BurningParrot stylesheet
No functional change. Homepage visual should not change. WebAuthn
passwordless login visual should not change. OIDC provider buttons are
still colored in homepage and login page.

Notes:
- Introduce new lib to share login styles. As a result, shared classes
are renamed.
- `to-snake-case` function is now private (with dash prefix).
- `.colored-button` class is only output in homepage styles. Since
tlp-swatch-colors are imported everywhere, previously this class was
being output a lot of times, but it's only needed in the homepage.
Sharing the mixin is ok since it does not produce output CSS unless it's
used.

Part of story #33005 login with only my passkey

Change-Id: I9dcb086f1d3914be6d7f2bb05322e94383e4cf6b