chore(jira): Extract export of an issue as XML in a dedicated object.
This is part of story #32788 Introduce the Jira project import with a solo tracker

No functional change expected here.
This object will be reused in both multi and mono tracker project
import. An issue is exported in XML in the same way.

This object is tested through export high level unit tests.

Change-Id: I74720fba890e9d8e971140365f088127c8ecc1d5