feat: Add post controller to switch passwordless only
Part of story #33005 login with only my passkey

This commit adds the controller for the switch button. This controller
will switch passwordless login only on/off.

*Testing:*

By clicking on the switch button, a form is submitted, you are
redirected on the same page with a feedback error as the feature is not
implemented

Change-Id: Id16e4fc682c2050ad34aa890b58e8b6cc5129e48