feat: add column passwordless_only in table user
Part of story #33005 login with only my passkey

Add a column in table use (default false) to save state of user
preference on passwordless only login. Add also method to dao and
UserManager to get and change this state.

Also use these methods in switch form and user preferences controller to
respectively switch and show state.

*Testing:*

On /plugins/webauthn/account, you can switch the state by clicking on
the switch button, the state is keeped when refreshing the page.

Change-Id: I3b7e5922c2ec0f8e6ca13fab5ea83c6d70ed8bdf