feat: Empty state when no Kanban for project
Go to a project without any Kanban (but with A.D activated). Go to the
/projects/<name>/kanban URL. You should have an empty state asking you
to come later (if you are project admin, you have a non-working button
to create a kanban instead).

Modal to create a Kanban will be added in a dedicated changeset.

Part of story #33646: access Kanban in a dedicated service

Change-Id: Ie2be0883e0bc8c82e84073158fa5b259441e8178