feat: switch off passwordless only state when deleting last passkey
Part of story #33005 login with only my passkey

If a user has enable passwordless only authentication and delete the
last registered passkey, the user is blocked outside of Tuleap. To avoid
this situation, the state is set to false then the user has no more
registered passkey

*Testing:*

Register a passkey, enable passwordless only, delete the key: the state
should be to false. As you can't see the switch button when you have no
key, you can check in database user.passwordless_only is set to 0.

Change-Id: I5fea84b0f9d37f5d590514d49518d69031b11079