Keep using threads even when running Vitest in coverage mode
This resolves the issue encountered when running the tests with coverage
in @tuleap/plugin-pullrequest-overview because it makes we do not end up
with the same custom element defined multiple times in the environment.

Note that we have disabled the threads in the past to workaround some
semi-random hangs when running the tests with coverage. Since the
situation is likely to have improved since the last time we checked
it is worth to revisit this (and also I really do not want to track in
JSDOM how to clean up the CE registry or to mock everything).

Part of request #33640 hybrids: 8.1.8 -> 8.2.2

Change-Id: I016b4fe09cbf0b1ea8b7e629ed0530d5598bc923