feat: Introduce feature flag
We've heard that some people are wary of changes due to split of A.D
into backlog and kanban. In order to bring a safety net to their support
team, here is a feature flag to deactivate the split.

If flag is activated (with a list of project ids) then those project
won't have splitted kanban and direct access to kanban homepage will be
redirected to legacy A.D homepage.

Example:
tuleap config-set feature_flag_temporarily_deactivate_splitted_kanban_for_project 123,124

This flag is only here to prevent big screams #REGRESSIONS 😱
Change will happen, it is just a matter of time and the flag allows
temporisation.

Part of story #33646: access Kanban in a dedicated service

Change-Id: I0b6c4767d3c0f71a2946fb06d0e5278f22dc23c5