chore: Add all missing move-artifact-action unit tests
part of request #33651 move-artifact-action app should be properly unit tested

I would like to apologize in advance for the size of this contribution.
I had to migrate the while store to TypeScript in the middle of the
writing because it made MoveModal impossible to test.

note: I'll migrate index.js and modal.js to TypeScript in a dedicated
contribution.

-- No functional change expectde

Change-Id: I0aa19e58b33d346433d56bc59b84f4e71ac074f8