fix: reset is_move_possible when another tracker is selected
part of request #33651 move-artifact-action app should be properly unit tested

When you select a tracker in which no field can be migrated, then the
[Confirm] button is disabled, and an error saying "the artifact cannot
be moved" is displayed.

Now that the bug making it impossible to change the tracker or the
project if the dry run step was run has been accidentally fixed in a
previous commit, the button is never reactivated, and the error message
is always shown.

How to test:
- Select a tracker in which no fields of your artifact can be moved
- Click on [Move artifact]
--> The modal says "the artifact cannot be moved"
--> The [Confirm] button is disabled

- Select another tracker in the current project or in another one
- Click on [Move artifact]
--> The error "the artifact cannot be moved" is no longer displayed
--> The [Confirm] button can be clicked

Change-Id: I2824d7d7e7c4a0e78f61a2c512a0656003e51d1a