feat: Introduce Kanban service
Kanban has now its dedicated service \o/

At upgrade, every projects using A.D service will get Kanban service.

Creating a project from Kanban template will activate Kanban service
instead of A.D service.

If both A.D and Kanban are used:
* A.D is displayed in the sidebar
* Kanban is displayed in the sidebar
* Kanban service leads to the Kanban homepage (no need to hack the url anymore)
* User can access a given Kanban via A.D or Kanban homepages
* The breadcrumb in a given Kanban targets Kanban homepage (in lieu of
  A.D homepage).
* Creating a Kanban from A.D redirects to Kanban homepage

If A.D is used, Kanban is used, but siteadmin decides to deactivate the
split for this project:
* Kanban is not anymore displayed in the sidebar
* The breadcrumb in a given Kanban leads to A.D homepage (even if
labeled Kanban)
* Creating a Kanban from A.D redirects to A.D homepage (see
implementation notes)

If A.D is not used and Kanban is used:
* A.D is not anymore displayed in the sidebar
* We can access to all feature of Kanban (widgets, homepage, …)

If A.D is not used, Kanban is used, but siteadmin decides to deactivate the
split for this project,
Or if both A.D and Kanban are not used:
* We cannot browse to access Kanban
* Direct access to Kanban homepage leads to NotFound
* Direct access to a given Kanban leads to NotFound

Implementation notes
--------------------

* Breadcrumbs always targets Kanban homepage. If no splitted kanban,
then user is automatically redirected to legacy homepage. This can
produce double redirection (Create kanban -> kanban homepage -> A.D
homepage) but since the splitted kanban is the new king in town, there
is no need to complexify the redirection code in CreateKanbanController,
for some behavior that will be removed later.
* No specific breadcrumbs are implemented in Kanban homepage. Waiting for UXD feedback about this.
* PluginWithService changes: 'Cause we need a little controversy^Wconsistency. PluginWithService is here to allow developers to understand what hooks to implement are related to service management. This is the case for serviceEnableForXmlImportRetriever (check that a plugin provides kanban service), it should be part of this interface. The same for `servicesAllowedForProject` (but is already implemented by Plugin class).

Part of story #33646: access Kanban in a dedicated service

Change-Id: I2f100ad142336fa8e53ee5cf451aa9b9a7cc300a